big initial commit with functional app
The couchapp is working (thanks to gdamjan's jquery code in my-couch.js)
but needs a lot of fixes to be deployed and used in production.
The style isn't added at all and some images and icons are also missing.
They will be added later on when the app is fully funcgtional.